+++ src/main/java/jinbok/culture/comment/service/CommentService.java
@@ -4,7 +4,10 @@ import jinbok.culture.comment.domain.Comment;
 import jinbok.culture.comment.dto.CommentRequest;
 import jinbok.culture.comment.dto.CommentResponse;
 import jinbok.culture.comment.repository.CommentRepository;
+import jinbok.culture.exception.RestApiException;
+import jinbok.culture.exception.code.CommentErrorCode;
 import jinbok.culture.user.domain.User;
+import jinbok.culture.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,12 +21,13 @@ import java.util.stream.Collectors;
 public class CommentService {
 
     private final CommentRepository commentRepository;
+    private final UserService userService;
 
-    public CommentResponse createComment(Long boardId, CommentRequest commentRequest, Object object) {
+    public CommentResponse createComment(Long boardId, CommentRequest commentRequest, Long userId) {
 
         Comment comment = Comment.builder()
                 .content(commentRequest.content())
-                .user((User) object)
+                .user(userService.getUser(userId))
                 .boardId(boardId)
                 .build();
 
@@ -33,21 +37,23 @@ public class CommentService {
     }
 
     public List<CommentResponse> findAllCommentsByBoardId(Long boardId) {
-        List<Comment> comment = commentRepository.findAllByBoardId(boardId);
-        return comment.stream().
-                map(CommentResponse::toCommentResponse).
-                collect(Collectors.toList());
+        List<Comment> comment = commentRepository.findAllByBoardId(boardId)
+                .orElseThrow(() -> new RestApiException(CommentErrorCode.INVALID_BOARD));
+
+        return comment.stream()
+                .map(CommentResponse::toCommentResponse)
+                .collect(Collectors.toList());
     }
 
-    public CommentResponse updateComment(Long boardId, Long commentId, CommentRequest commentRequest, Object object) {
-        Comment comment = commentRepository.findById(commentId).orElseThrow();
+    public CommentResponse updateComment(Long boardId, Long commentId, CommentRequest commentRequest, Long userId) {
+        Comment comment = findCommentById(commentId);
 
         if (!comment.getBoardId().equals(boardId)) {
-            throw new IllegalArgumentException("댓글이 옳지 않은 게시글에 존재");
+            throw new RestApiException(CommentErrorCode.INVALID_BOARD);
         }
 
-        if (!comment.getUser().getId().equals(((User) object).getId())) {
-            throw new IllegalArgumentException("옳지 않은 유저가 접근");
+        if (!comment.getUser().getId().equals(userId)) {
+            throw new RestApiException(CommentErrorCode.INVALID_USER);
         }
 
         comment.updateComment(commentRequest);
@@ -55,16 +61,26 @@ public class CommentService {
         return CommentResponse.toCommentResponse(comment);
     }
 
-    public CommentResponse deleteComment(Long boardId, Long commentId) {
-        Comment comment = commentRepository.findById(commentId).orElseThrow();
+    public CommentResponse deleteComment(Long boardId, Long commentId, Long userId) {
+
+        Comment comment = findCommentById(commentId);
 
         if (!comment.getBoardId().equals(boardId)) {
-            throw new IllegalArgumentException("댓글이 옳지 않은 게시글에 존재");
+            throw new RestApiException(CommentErrorCode.INVALID_BOARD);
+        }
+
+        if (!comment.getUser().getId().equals(userId)) {
+            throw new RestApiException(CommentErrorCode.INVALID_USER);
         }
 
         commentRepository.delete(comment);
 
         return CommentResponse.toCommentResponse(comment);
+    }
+
+    private Comment findCommentById(Long commentId) {
+        return commentRepository.findById(commentId)
+                .orElseThrow(() -> new RestApiException(CommentErrorCode.INVALID_COMMENT));
     }
 
 }
